Fixed deviceIndex bug and simplified code
The deviceIndex bug was where you don't have to be connected to a Bluetooth device to check its address, so I added in an actual isConnected function. The code for basic stats was totally useless so I decided to ditch it, too. Also changed/added the info and warning logs.